chore: rename utils to func
