+++ packages/element-lib/src/sidetree-v2/func.js
